clean up API and add unit tests/README

+++ library/src/main/java/org/michaelevans/uber/services/UberService.java
@@ -38,5 +38,5 @@ public interface UberService {
     UberPrices getPricingEstimate(@Query("start_latitude") double start_latitude, @Query("start_longitude") double start_longitude, @Query("end_latitude") double end_latitude, @Query("end_longitude") double end_longitude);
 
     @GET("/v1/estimates/time")
-    UberTimes getTimeEstimate(@Query("start_latitude") double start_latitude, @Query("start_longitude") float start_longitude, @Query("customer_uuid") String customerUUID, @Query("product_id") String product_id);
+    UberTimes getTimeEstimate(@Query("start_latitude") double start_latitude, @Query("start_longitude") double start_longitude, @Query("customer_uuid") String customerUUID, @Query("product_id") String product_id);
 }
